+++ server/public/style.css
@@ -1,0 +1,4 @@
+.header{
+  display: flex;
+  justify-content: space-between;
+}
